fix:simulacion con threads
